[WiP] Split up code into files and added more tests.
